requirements.txt
pandas>=2.2,<3.1
requests>=2.32,<3
psycopg2-binary>=2.9,<3
SQLAlchemy>=2.0,<3
boto3>=1.37,<2
python-dotenv>=1.0,<2
pyarrow>=18,<20
matplotlib>=3.9,<4
seaborn>=0.13,<0.14
